backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
psycopg2-binary==2.9.10
httpx==0.28.0
python-dotenv==1.0.1
asyncpg==0.29.0
alembic==1.13.1
openai>=1.0.0
mcp>=0.1.0